requirements.txt
albumentations==1.3.1
numpy==1.26.2
opencv_python_headless==4.8.1.78
pandas==2.1.3
PyYAML==6.0.1
PyYAML==6.0.1
scikit_learn==1.3.2
torch==2.1.1
tqdm==4.66.1
